fix: Separate router and worker runtime entrypoints for reloaded lib
