+++ experiments/billable/worker-configuration.d.ts
@@ -1,8 +1,8 @@
 // Generated by Wrangler by running `wrangler types`
 
 interface Env {
-	RESEND_API_KEY: "123";
-	APP_URL: "http://localhost:5173";
+	APP_URL: string;
+	RESEND_API_KEY: string;
 	SESSION_DO: DurableObjectNamespace<import("./src/worker").SessionDO>;
 	R2: R2Bucket;
 	DB: D1Database;
