+++ reloaded/src/runtime/entries/router.tsx
@@ -1,0 +1,2 @@
+export * from "../lib/router";
+export * from "../lib/links";
