+++ reloaded/src/runtime/entries/worker.tsx
@@ -1,0 +1,8 @@
+import "react";
+import "react-dom/server.edge";
+import "react-server-dom-webpack/server.edge";
+import "react-server-dom-webpack/client.edge";
+
+export * from "../worker";
+export * from "../db";
+export * from "../error";
