--- reloaded/src/worker/index.tsx
@@ -1,5 +1,0 @@
-export * from "./worker";
-export * from "./lib/router";
-export * from "./lib/links";
-export * from "./db";
-export * from "./error";
